Rework the MongoDB driver modules
