+++ modules/driver-mongo-mapping/driver-mongo-mapping-jackson/build.gradle.kts
@@ -2,7 +2,7 @@ plugins {
     id("datamaintain.conventions.kotlin")
     id("datamaintain.conventions.publishing")
     id("datamaintain.conventions.driver")
-    id("datamaintain.conventions.driver.mongo")
+    id("datamaintain.conventions.driver.mongo-mapping")
 }
 
 dependencies {
